requirements.production.txt
# Production Requirements for Cooplink
# These are the exact versions tested for production deployment

# Core Django Framework
Django==4.2.7
django-environ==0.11.2
python-decouple==3.8

# Database
psycopg2-binary==2.9.9
django-redis==5.4.0
redis==5.0.1

# API Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.0

# Security & Authentication
django-cors-headers==4.3.1
django-ratelimit==4.1.0
cryptography==41.0.7

# File Storage & Static Files
django-storages==1.14.2
boto3==1.34.0
botocore==1.34.0
whitenoise==6.6.0

# Image Processing
Pillow==10.1.0

# Admin Interface
django-unfold==0.18.0

# Markdown Support
django-markdownx==4.0.2
markdown==3.5.1

# Logging & Monitoring
structlog==23.2.0
sentry-sdk==1.38.0

# HTTP Client
requests==2.31.0
urllib3==2.0.7

# Date/Time Utilities
python-dateutil==2.8.2

# Environment & Configuration
python-dotenv==1.0.0

# Production Server (Optional - choose one)
gunicorn==21.2.0
# uWSGI==2.0.22

# Performance & Caching
django-compressor==4.4

# Database Optimizations
django-cachalot==2.6.1

# Email Backend (Optional)
django-ses==3.5.0

# Task Queue (Optional)
celery==5.3.4
kombu==5.3.4

# Health Checks
django-health-check==3.17.0

# Security Headers
django-security==0.17.0

# Rate Limiting
django-ratelimit==4.1.0

# API Documentation
drf-spectacular==0.27.0

# Timezone Support
pytz==2023.3

# JSON Web Tokens
PyJWT==2.8.0

# UUID Support
uuid==1.30

# Decimal Support
decimal==1.70

# Password Hashing
argon2-cffi==23.1.0
bcrypt==4.1.2

# Development Tools (for deployment script)
# Remove these in production if not needed
coverage==7.3.2
factory-boy==3.3.0

# Production Monitoring
django-extensions==3.2.3

# Memory Profiling (Optional)
memory-profiler==0.61.0

# System Monitoring
psutil==5.9.6
